feat: add check for files/directories

+++ src/file/fileDeleter.js
@@ -1,7 +1,8 @@
-import {unlink, access} from 'fs/promises';
+import {unlink, stat} from 'fs/promises';
 import {resolve} from 'path';
 
 export const deleteFile = async (curDir, filePath) => {
-    await access(resolve(curDir,filePath)); /* Check for file existence/accessibility */
+    /* Check that path leads to file, not to directory */
+    if (!(await stat(resolve(curDir, filePath))).isFile()) throw Error();
     await unlink(resolve(curDir, filePath));
 };
